Include lecture attendance on participant page
Leaders can see whether or not a participant has attended lectures.
Give activity chairs the ability to mark any participant as having
attended lectures.

When it's allowed to do so, participants can mark themselves as having
attended lectures. This should only be permitted for a small window
during the start of lectures.

Once Winter School is underway, we'll display a warning to any
participant who has not attended lectures.