Fix editing leaders on an existing trip
The leader selection widget will be invalidated whenever any selected
leader is not able to lead the trip. However, we performed this
invalidation before actually checking if everyone can lead the trip.
That is, `canLead` was `undefined` and thus falsey - meaning that the
most recently-added leader was always erroneously marked as not being
able to lead the trip.

Fix this by always having the validator check its prerequisites.