Show notification icon for pending applications
Currently, the activity chairs have no way of knowing how many pending
applications there are. Give them a small count of outstanding
notifications as part of the menu.

In order to support this, we need to completely refactor how we identify
pending applications. This is well-timed, as the application-processing
code was hastily constructed and in need of cleanup.