Improve clarity on lottery driving
If a participant goes to delete their car information, point them
towards the lottery preferences page instead (which is what they
probably want to indicate that they're not driving).

If they delete their car altogether, but are still marked as a driver in
the lottery, then prompt them to change their preferences (or add car
information).