Resolve 500 on GET delete, update drop logic
Deleting database objects on a GET is a solidly bad idea. Thankfully,
Django just wants to render a template with the default DeleteView.
These templates are not implemented, so we should instead redirect to
the appropriate pages where POST-generating buttons are present instead.

Also fix some uses of model.id to instead use model.pk