Clean up logic around chairs and admins
Restores admin ability to see "activity chair" functions on trips
without a formal activity chair.

This also makes the intent of the `is_chair()` function very clear. We
were previously using it as a "chairs or admins allowed" permission
check, but also expecting it to act as an authoritative source on
whether a given activity had a chair.

This leaves a gap in coverage - some activities need to be
administrated, but you cannot assign ratings or view applications for
the activity. Make sure that we cover all cases:

1. Admins and activity chairs (if any) and admins need access
2. Just activity chairs need access
3. Admins and chairs need access, but only if activity has chairs