Display when waiver current, membership missing
"Mini trips" are those that require active waivers, but do not strictly
require an active MITOC membership.

For simplicity, we'd previously considered a member with an active
waiver but unpaid membership to simply be "missing" their membership.
Now, we report when their status is missing a waiver.

NOTE: Due to the way that we currently handle membership processing, we
don't actually have entries in our database for a given person with a
waiver until they sign up for a membership. This change is still useful
because:

1. We'll be changing our membership-processing script soon
2. There are some participants whose memberships have lapsed, but can
   still have up-to-date waivers