Fix rating for old chairs, better UX updating recs
If you're an activity chair that previously gave a rating suggestion to
a given participant, you'll experience a very strange UX when reviewing
their application:

You'll first be prompted to make a rating for the participant. After you
select 'Is a recommendation' and submit your recommendation, nothing
shows up. The reason this occurs is that we were updating old
recommendations from a previous application, rather than creating a new
application.

This commit fixes that bug, and also adds some better UX to make it
clear what's going on. You can update your existing recommendation, and
the button should make that clear.