Move the attendance div to the top on all screens
At tonight's lectures, participants will be able to mark their
attendance. Because the current grid system pushes this div all the way
to the bottom of the screen, there may be some confusion. Use Bootstrap
classes & a redundant div to move content all the way to top of the
viewport.