Add a route to fetch other verified emails
Provide a email-checking that can be accessed with a GET (no need for
CORS). This allows external services to query the trips database for
information about members. To verify their request, all we need is a
signature accompanying the email address.

JWT is pretty overkill for this use case, so just use HMAC (supported
by Python builtins) to verify a signature.