Allow specifying parent/guardian on waiver
MITOC members under the age of 18 must give a parent or guardian when
completing a waiver. This gives the ability for participants to specify
who that guardian is.

We're basically building a PowerForm. Users specify roles themselves,
which then creates an envelope. On envelope creation, the API is
notified and we upload the participant's membership.

The only reason we simply can't link participants to the PowerForm is
that the Docusign Connect API doesn't come with MIT's account: the only
way to be notified when an envelope is complete is for us to actually
create the envelope.