Just use dates on rentals
The `checkedout` column is technically a full datetime, but it just uses
the current day at 05:00 UTC (i.e. midnight on the day the item was
checked out). Since the app only cares about the _date_ when an item was
rented, use that instead.

This also eliminates a bug around timezone-aware and timezone-naive
datetimes.