Consistently format response on email lookup
In cases where the MITOCer is not previously known to us, we just return
the email that was given. However, this means that consumers of the API
can expect a `primary` email _some_ of the time, but not consistently.

Change the API response so that the primary is always given in addition
to all known emails.