Allow leaders to sign up as participants
For Circuses, leaders have the ability to sign themselves up as leaders.
However, in some cases, a person may not be able to act as a leader
(perhaps they have a 3-season hiking rating, and it's a Winter Circus).
They may just not want to lead trips. In these cases, they should be
able to sign up as a participant instead.