Hide signup if already leading the trip
This is an existing bug (i.e. it predates the template re-arrangement).
For leaders who signed up for the trip as leaders, they will still see
the option to sign up.

Just as an existing leader should not see leader signup, the same person
cannot lead a trip that they are a normal participant on.