Support having a WIMP for a single trip
In Winter School, we have a WIMP for all trips in a given weekend. In
the rest of the year, this model doesn't translate well. Individual
trips may want to have a single person who is responsible for ensuring
that everybody gets back safely. This person should be easily specified
by the leaders, and should not be granted access to other trips'
information.