Don't show WIMP toolbar to anonymous users
As an anonymous user, the Participant object passed with each request is
`None`. When a trip has no WIMP, `trip.wimp` is also equal to `None`.
Failing to consider this equality will result in showing the WIMP
toolbar (but none of the underlying participant information!) to
users who are not logged in.

To correct this, we should ensure that the WIMP for each trip is not
`None` _before_ comparing it to the viewing participant object.

With thanks to Kyle Barron for the bug report:
https://github.com/DavidCain/mitoc-trips/issues/15