Move "Participant from user" to be a model method
Previously, we'd only translated from a User object to its corresponding
Participant in middleware. Accordingly, it was acceptable to just keep
it as a one-off metod. However, we'll soon want to perform this
database-spanning translation in a variety of other places. For ease,
make it a method of the Participant model class.