Framework to cache membership & waiver data
The current membership database is hosted on sql.mit.edu and is quite
sluggish. By caching membership & waiver data in the trips database, we
can significantly decrease load times and facilitate simple checks of
"is the participant an active member."

We plan on enforcing current membership/waiver when a participant goes
to sign up. Rather than checking the sql.mit.edu database (and waiting
for a response) each time, we could ask just once and cache that
information for future use.

In most cases, it doesn't actually matter _when_ the participant's
waiver/membership expires, so long as that date is in the future. For
example, we may cache a waiver with an expiration date 6 months in the
future. If that participant signs another waiver (expiring in 12
months), it doesn't matter if we don't cache it, because the old waiver
will still be valid for another 6 months.