Fix Angular bugs (scope.leaders is undefined)
When we use redirection-based decorators on JSON routes, we can return
responses with the wrong mimetype, which cause issues in Angular code.
An HTML redirect response is not viewed as failure, and the Angular then
tries to access the data as if it was a JSON Response.

A participant with lapsed data would previously trigger console errors
when viewing a trip, since the JSON routes used in Angular directives
return HTML instead of a JSON response with a 400 error code.