Force a refresh after trip modification
This solves a longstanding annoyance where the static content on the
Trip tab would become dated after making changes on the Admin tab.
Now, if you switch back to the Trip tab, the page will do a hard refresh
automatically.

This could cause some problems if a leader had partial (but unsaved)
changes existing in the Admin tab after they've already submitted one
set of changes. I think this is a rare enough occurrence that this is
a net UX improvement.

The ideal fix, going forward, would be to update just the table of
participants & waiting list, leaving the rest of the table in-place.