Populate the geardb with expected tables
Because the gear database is not managed by this project, the tables
aren't defined in models and thus do not get created as part of unit
tests.

Use a really minimalist, pure SQL method of copying the expected schema
directly into the database so we can use it in tests.