Query for updated waiver status
When a participant signs a waiver, we currently show them a one-time
message that reads "we're processing your waiver," but we also
(statically) show the old waiver's expiration date. It only takes a
couple seconds for DocuSign to POST to our processing pipeline, but some
participants don't know to refresh the page and check for updates.

This UX is slightly confusing, and has led some participants to sign two
waivers in a fairly short period of time.

As an improvement, we should instead show that we're processing the
update, and repeatedly query the backend for an update.