Waiver update logic even if already active
If the waiver you had on file is currently active, but expiring in a few
days, you might choose to sign another waiver right away. The delayed
waiver display logic should also handle this case, and not just the case
of participants signing a waiver because the old one expired.