Always give links when unable to attend trip
When telling participants why they cannot sign up for a given trip, we
should always give them some easy links to resolve the situation.

In almost all cases, the membership check should provide a clear
explanation as to why a participant cannot go on the trip. However, it
is harmless to add a few links in the description to make things easier.

Thanks to Chris Blatchley for pointing this out.