Change weighting on past participation
In previous Winter Schools, participants were effectively penalized for
being active early on in Winter School. Participants who had not been on
any trips were always given their choice of trips before others.

This was an intentional design decision to ensure that everyone in the
community had a chance to go on trips. However, recent Winter Schools
have had a surplus of trips, and it's quite easy for all participants to
go on at least one trip if they want to.

An unintended consequence of this design decision is that experienced
Winter School participants would be placed behind complete beginners on
late-season trips (where experience is especially valuable).

Since we: (1) have plenty of open trips (2) don't want to encourage
people to sit out trips in order to game the system, we should just do
away with the trip-based ranking altogether.

We'll still give active leaders a bonus for having led a lot of trips
recently. However, we should only consider recent activity when granting
this bonus.