Boost MIT students/affiliattes on trip lottery
When running a lottery for a single trip, we should always give MIT
students a significant advantage over non-affiliates.

We've always done this automatically in the Winter School lottery, and
leaders have explicitly promoted MIT students on oversubscribed single
trips. However, it's a significant improvement to automatically
give MIT students an advantage in single trip lotteries.

All else held equal, this means that an MIT undergrad would be placed on
a trip before a non-undergrad ~75% of the time, an MIT grad would beat
out a non-affiliate ~68% of the time, and an MIT affiliate would win out
~60% of the time.