For simplicity, only consider WS'ers with signups
We now have over 3,000 participants in the system. Many of them have
long since left MITOC, and showing 2,000 instances of "so and so did not
choose any trips this week" is not useful in the log. Only consider
people who've actually signed up.

Additionally, since the Winter School lottery logic makes a number of
queries (it's not optimized for speed), this should speed things up
notably.