Remove old, unused signals
We used to send people emails informing them that they were bumped off
the waitlist and were placed on the trip. However, we stopped doing this
due to erroneous emails being sent out.

Now, this same signal (even though it is effectively disabled) is
causing problems with trip deletion. We should just delete it at this
point.