Start using FactoryBoy
We'll use FactoryBoy to create actual database objects so that the SQL
layer isn't always being mocked away in unit tests. (We'll continue to
make heavy use of mocking, but not at the expense of test gaps on
database queries).

Start by declaring the dependency, defining a few key factories for use
in tests, then finally writing a basic test that uses the new factories.