Move group creation into a migration
Rather than dynamically creating Groups on app load, we should instead
just run a data migration.

This cleans up an ugly old pattern, but also frees up Groups to be
referenced in unit tests. Specifically, a number of key Participant
signals expect groups to exist in the database.

This removes a hack that was needed to handle the runtime's assumption
of data being present in the database. Since the group-creating code was
executed on every import of the `ws` package, we needed to wrap in it a
try/except statement. This way, `manage.py` commands would not
immediately exit from the resulting error(s) due to a not-yet-migrated
database.

(In fact, this setup was necessary to run migrations on a blank database
in the first place).