Defer most assets for speed
Most views of the site will be with cached assets. However, to speed up
initial visitor experience (or the first visit after a deploy), we
should defer all non-essential blocking stylesheets/JS.

I choose to keep the main CSS bundle blocking, to eliminate the unsavory
"flicker" when loading the DOM before Bootstrap is initialized. Raven
is also blocking (for now) since we need to ensure that it's executed
before the `script` tag setting user context is executed.