Defer connection with Google Sheets
Any management command that inadvertently loads the `member_sheets`
module will attempt to perform a handshake with Google Sheets to verify
credentials and initalize a session.

This is obviously unnecessary for normal management commands. However,
it also introduces a potential for production bugs: if the credentials
file is only readable by a privileged user, running management commands
as a different user might raise exceptions when importing the module.