Add method to update geardb with new affiliation
Since the gear database has long lacked accurate affiliation data for
many MITOC members, we should add a method to sync affiliations from the
trips database to the gear database.

This method should be run once for all current participants, to bring
the databases into sync. Going forward, we should run it periodically
for all participants, to make sure that the gear database has the most
up-to-date information.