Update the gear database after profile changes
Whenever participants update their profile, we should take that
opportunity to ensure the gear database is up-to-date.

Even in the case when the affiliation is unchanged on the trips
database, we should make an update. For example, a MIT student could
have accidentally paid for a non-affiliate membership, which updates
their account. When they assert through mitoc-trips that they are indeed
a MIT student, we should overwrite their last-submitted membership.