Format phone numbers
For a long time, we've been displaying phone numbers in their raw E164
format, without any country-specific styling. Correct this. The vast
majority of numbers are US numbers, which we can display in standard
national format: (000) 000-0000. International numbers should be
displayed in their country-specific formatting, with the country code
leading.