Add a link explaining Markdown syntax
Most leaders are wise to the fact that trips can be described using
Markdown (either through experience, or via the example trip
description). However, make it explicit for others who may not be aware!

Some leaders have been including images in their trip descriptions. This
is not always a great practice, since it imposes needless bandwidth
constraints for users accessing the site in remote environments.
Furthermore, we should be very sure that the content owners allow MITOC
hosting these images. Remind users that they must have permission to use
images in their description.

Lastly, the content policy will forbid images hosted over HTTP. Make a
note of this, if briefly.

Credit to Kyle Barron for the suggestion on linking this Markdown syntax
guide!