Fix JSON leader view being publicly accessible
We have an API route that returns the name and profile images for all
MITOC leaders. This is a public route, currently used by the MITOC
static site. For example, each of the following pages lists the current
leaders for the activity:

- https://mitoc.mit.edu/#become-3-season-leader
- https://mitoc.mit.edu/#become-paddling-leader
- https://mitoc.mit.edu/#become-ws-leader
- https://mitoc.mit.edu/#become-climbing-leader

Each page calls this API route, which is currently giving a 302 and
redirecting to the login page (which prevents the leader listing
functionality from working properly).

Functionality was broken in a past refactoring of class inheritance.
This commit restores old functionality.