Give participants guidance around accounts
Every Winter School, many participants are confused that paying for
their dues doesn't automatically create an account for them. They input
their "usual" password (which is frightening), and are surprised when
it doesn't work.

This commit attempts to add some clarity on the account flow pages as to
what an account actually _is_ and how one should go about creating one.

This also removes i18n from our template (this project is always in
English, we won't internationalize), as well as some dated CSS.