Reduce the query count on trip medical pages
It takes a while for the WIMP view to fully load. Perform some SQL
optimizations to drive the total query count down.

- When checking permissions for viewing the page, start with more likely
  checks that also don't need to hit the database (checking against
  fetched groups, for instance)
- Avoid re-fetching the same trip object in the DetailView
- JOIN on emergency contact information

These optimizations bring the trips/<pk>/medical view down to around 4
SQL queries for the average viewer. On trips/medical/ (WIMP information
sheet), we still have an intentionally inefficient system which performs
those same 4 queries for each trip. We can fix that up later, but it's
pretty speedy for now (and the page is rarely viewed).