Add even more to lottery output logs
We have an interest in analyzing the output of the lottery runs so that
we can evaluate fairness, identify Pareto-efficient signup swaps (if
any), and ensure MITOC is serving its student body.

Adding a JSON-parseable dictionary of some key stats for each
participant's results will make it easier for us to generate lottery
statistics.