Implement a hacky listing of form errors
When form validation fails, Djangular is supposed to load the
server-provided errors into the appropriate widget for each form
element. However, the way that we do nested forms for participants
doesn't work nicely with this. Front-end validation catches 99% of all
errors that might happen, but there are a few extra cases that might
sneak by.

Until we can switch to a better frontend architecture, just manually
render server-generated errors if there's any trouble.

This also catched a weird bug where the minimum year validation is not
applying to the car year on the car form.