Rework how we handle applications
The concept of restricting applications to those "just in a given year"
is not terribly helpful for most activities. Winter School is the only
activity where applications are submitted in a given calendar year and
always dealt with in that same year. In other activities, participants
could submit an application in December, only for activity chairs to see
the same application disappear in January!

This commit fixes that bug. Once an application is submitted, it remains
in the "pending" queue until activity chairs deal with it. At present,
the only ways to clear an application from the queue are to:

1. Grant the participant a rating
   - This is the usual case, but sometimes prospective leaders aren't
   actually given a rating
2. Grant a rating, then immediately delete it
   - This amounts to never giving them an actual rating, but it's an
   ugly workaround, and the record makes it look like they were actually
   a leader for a while.
3. Hard delete the application
   -  This definitely clears the application (and allows the participant
   to submit another at a later time). However, it has the obvious
   downside of data loss.

I'll be fixing the above issues by implementing the ability for chairs
to "dismiss" an application.

In the meantime, though, this commit:

- Significantly improves test coverage
- Simplifies our existing API
- Adds comments
- Reduces SQL queries needed on each page