Use `isort` to enforce import ordering
`black` handles almost all aspects of code styling. `isort` can be used
to make sure that imports adhere to PEP 8 recommendations. However, I
want to be sure that `isort` does not apply any style changes that
`black` will then want to change!

The author of `black` went to great lengths to ensure that the tool has
the same output as `isort`'s "mode 3 + trailing commas." I have a lot of
interest in using both tools with as minimal configuration as possible.
The included `.isort.cfg` makes sure that it doesn't interfere with
`black`.