Send a weekly blast of trips (pre-formatting)
This adds the ability to add a server-generated description of all
upcoming trips, sending an email to a Mailman mailing list
(mitoc-trip-announce@mit.edu). At present, we've not yet configured it
as a recurring Celery job, but we will do that after testing.

The HTML templates have one critical flaw, though, which will be
addressed in a successive commit: They have not been CSS-inlined.