Use Vue with full runtime compiler
If we wish to inject Vue for use with DOM templates (which is a strict
requirement for building Vue components in the backend), we'll need the
full runtime compiler available.

Vue-CLI does not provide this by default.

Also, add a root `div` in our base template, much like Angular listens
to the `body` tag (because of Vue's virtual DOM, it cannot listen on
`body`).