Create a membership status component
We conditionally display (and stylize) each MITOC member's status.
Encapsulate this behavior in a single, testable component. It will prove
useful in a top-level component that obtains and displays information
about a given user's membership.

It need not be registered globally, since it will only be used by a
single component.