Add a component for membership status summary
When a given user is viewing their membership, they often ask questions
about the displayed status. Common questions include:
- Why is no membership found when I just paid/signed a waiver?
- Why is a waiver expired when I just signed one?

There are common answers to these questions, and some resolutions
require admin intervention. We provide a simple summary that hopefully
makes it more clear to people _why_ we display the information that we
have, and where they can go for help.