Replace Angular directive with Vue component
We've created a Vue component (MembershipStatus) that replaces the
Angular directive of the same name.

While we're in transition, both Vue and Angular will be watching the DOM
to see if they should apply changes to a given element. It's important
to not have any collisions on element names, so we transfer control to
Vue.