Auto-register base components
Rather than having to manually enumerate every possible component in
`main.ts`, explicitly register each component present in `/components`.
This makes a few key assumptions (leading to a simpler implementation
than what's given in the docs):

1. We will always use `export default` in each base component
2. All components will use `.vue`
3. *All* components in the root directory are intended for as base components

I expect #3 to change in the future, but #1 and #2 are rules I'll be
sticking to for the foreseeable future.