Cease using `dangerHover` directive
This directive was only used in one place - removing an email address.
This action isn't *that* dangerous anyway, since you can easily re-add
the same email address.

If we wish to re-implement this, the solution is to make a custom CSS
class for buttons that change their styling with `:hover`.

Steadily dismantling our old Angular directives will help ease the
transition to Vue.