Paginate past trips by default
We've always supported rendering *all* trips on a single page. The SQL
that fetches all this data is fairly optimized, and the DOM is not
excessively complicated. Still, we have years of past trips, with over
1,000 total. Accordingly, this approach can result in slow page renders
for mobile clients or anybody on a slower connection/device.

Offer a simple pagination-like approach to trip enumeration  By default,
only show trips in the past year (with the option to keep clicking for
additional years' backlog, yielding a fairly obvious URL scheme that can
be adapted for custom queries).

This should make for faster response times.

This change is the first view to have direct testing! In order to parse
HTML responses, we include BeautifulSoup as a dev dependency.