Create the concept of a trip 'program'
A trip program will soon replace the idea of an 'activity' altogether.

At present, we use a trip's activity in a number of ways:

- to restrict creation of trips to leaders certified by the activity
  chair (e.g. you need the WSC to consider you a leader in order to
  create WS trips)
- (sometimes) to describe the type of activity you'll be undertaking
  (hiking or boating are descriptive enough as-is, but "winter school"
  definitely is not)
- (sometimes) to describe the way the system handles trip admin.
  For example, WS trips run as part of a special multi-trip lottery
  with itinerary submission rules, lecture attendance requirements,
  ranking by participants, etc.

This adds a new (optional) field to trips. In successive commits, we'll
take apart the existing `activity` system altogether and replace it with
a program-based system.