Add coverage + coveralls for reporting
One of the long-time shortcomings of this repository is its lack of
comprehensive unit testing. I've been trying to address that in a couple
of ways:

- All new JavaScript code _must_ have complete test coverage.
  This rule is part of the current rewrite to VueJS
- Add bits of test coverage in bite-sized chunks as time goes on.

Ideally, I'd start by testing the most mission-critical code. In
practice, though, I'm starting by testing some of the simpler (read:
easy to test) parts of the infrastructure.

Over time, though, the total code coverage should be _increasing_, and
recording coverage as part of the build process will help measure that.