State that isort should have the pyproject extra
This is the poetry equivalent of `pip install isort[pyproject]`.

`isort` must be able to read configuration from `pyproject.toml` and it
requires the `toml` package to do this. Incidentally, `black` also
requires `toml`, so everything works wonderfully *when Black is
installed.*

However, if we were to abandon Black for some reason, that would remove
the `toml` subdependency, and `isort` would cease to be able to read
from `pyproject.toml`. Properly express this dependency!