Add a Makefile to configure build
As I've been expanding build tools (linters, code formatters, coverage
reporting, etc.) there has been an increase in the number of commands
one needs to know in order to interact the repository. For example,
three tools (isort, black, and pylint) all lint the source.

Additionally, the Vue rollout has a number of `npm run` commands that
are needed to invoke the test suite, run a local development server,
etc.

It's high time to make a simple Makefile that collects all this in one
place (and ensures that local development matches what's done on Travis).