Use Have I Been Pwned on login
We're already making use of the Have I Been Pwned API to make sure that
passwords aren't breached when users set a password for themselves (this
is done through a password validator).

Every time users log in, though, we briefly have access to their
password (before it's salted, then hashed). This is a good opportunity
to ensure that users have a secure password.