Log when password was last checked
It will help inform the rules on this password policy to log when
participants last had their password compared against Have I Been
Pwned's database.

We _could_ use `auth_user.last_login` to infer that information, but
it's possible that there was an API outage that prevented checking on a
given login.