Add some test coverage of local/prod settings
We should have test coverage of both the production and local setups!
It's particularly important to actually run the code in
`production_settings.py`, since local development will pretty much never
invoke it.

Additionally, `getenv()` allows the environment to be missing some
variables, but this should really be regarded as a critical error that
raises an exception and stops the worker from booting.