Make all pages display messages by default
Opting to not have a page display messages can make for some weird UX.
For example, if a user logs in, but is redirected to a page *without*
messages, the user can spend quite a bit of time on that page. Once
navigating to a page that *does* have messages, the user will then be
shown a "log successful" message (confusingly, long after logging in).

Similarly, separate views might generate the same message multiple
times, only for the message to be shown several times upon the first
navigation to a message-displaying page.

This will also lay the groundwork for alerting about important messages
on *every* page.