Improve the "internal server error" page
We should make efforts to make sure that all internal errors produce a
template that can be rendered and displayed to the user.

When a server error occurs, key middleware will not be functioning.
Namely, we will not have access to the request's user or participant,
nor the messages supplied by Django's messages framework.

Without the participant object, rendering the full menu is no longer
possible (nor do we have much useful to show).