Only log message once per request
Logging that we warned the user many times adds bulk to the logs.
Furthermore, it doesn't reflect the reality of how many times that user
actually sees a single message. If they're hitting other routes that
go through middleware (but do not load messages), they might not
actually see the messages.