Fix waitlist ordering bug
There has been a bug with the waitlist ordering, where a change in the
`NULLS FIRST`/`NULLS LAST` behavior across database versions &
implementations changed. We lacked test coverage, so the bug went
uncaught (until it was reported by attentive leaders).

Add some test coverage around waitlist ordering, which should prevent
this in the future, and give us some confidence that behavior is
unchanged when we upgrade to Django 2 and use the F-expressions that
come with Django 2.