Stop using Raven in JavaScript
Raven is the deprecated Sentry client for JavaScript. When we start
using Vue, we're going to be using the newer client. It would certainly
cause significant errors to have two different Sentry clients running at
once.

This change enables us to advance two goals separately:

1. Moving off of AngularJS in favor of Vue
    - dropping one more Angular-specific dependency is a good thing
2. Dropping the legacy Sentry client
    - we're going to use the newer client

At present, this does actually mean that we have no error reporting for
JavaScript. I'm okay with that, because:

- we haven't made any significant AngularJS changes in a long, long time
- Sentry has been extremely quiet (read: zero real errors) for JS
    - only reported errors are noise from browser extensions
- we're going to be re-instating a Sentry client soon