Expose an interface for users to bulk unsubscribe
A common complaint of MITOCers is that it's too difficult to unsubscribe
from Mailman lists. The link in each footer provides an interface to
*request* unsubscription, but it doesn't provide one-click unsubscribes
(i.e. a link with a token in it) that good mailing lists do.

To work around this shortcoming, plus the usual point of confusion where
users actually have no idea what lists they may be subscribed to, expose
a simple interface that can submit unsubscribe requests on their behalf.