Separate out legacy JavaScript
We're going to be slowly porting over the application from the legacy
AngularJS setup to our newer Vue setup. Intentionally allowing templates
to decline loading legacy JavaScript is the first meaningful step in
doing that rollover!

This will also lay the framework for letting some pages use the new
Bootstrap 4, while others stay on Bootstrap 3 to appease the legacy
dependencies.