Add a new trip type to each trip
`trip_type` is meant to be a descriptor that tells participants what
sort of activity will take place on the trip. It is distinctly separated
from the system of leader ratings and the programs in which leaders can
create trips. For instance, any trip can be described as 'backcountry skiing'
since we do not enforce that the leader has any rating-granted
ability to create backcountry skiing trip. Leaders *should* know that a
rating from the Winter Safety Committee is required to create skiing
trips, but that's up to leaders & activity chairs to enforce.

`trip_type` will eventually replace the meaning of `activity`, in the
context of "what will be doing on this trip?"

For now, it is an optional field that has no actual bearing on system
functionality.

We migrate all existing trips to have the correct `trip_type`, and will
soon work on surfacing this field to leaders.