Remove the leaderboard
As fun as the leaderboard was to build, it's impeding a few ongoing
initiatives. Namely:

1. Moving off of legacy AngularJS towards a Vue-based frontend
2. Changing the meaning of a trip's "activity" to remove tight coupling
   between leader ratings, trip descriptions, and trip programs.

I will likely replace this leaderboard with a newer, better, snazzier
version once we're on Vue and the trip data model has changed.

The backend route is still left in place at present, so users on a stale
client can still use the leaderboard (I'll remove the route at a later
date).