Add an API route that returns program leaders
The route that returns all leaders by activity is going to be removed!
The front end does not need to know all leaders for all activities when
creating a trip. Instead, it should only need to know the leaders who
can be used for a single trip: After setting the trip's program, we can
fetch a list of valid leaders. This should all happen in the background
before getting to the leader multi-select.

This should provide a small performance boost to the front end. Most
importantly, it will aid in future modifications to leader eligibility.
For instance, we may wish to have a separate enumeration of leaders able
to create trips for the MITOC Rock Program (instead of just letting all
leaders with a rating lead trips). By asking the back end "who can lead
MRP trips?" we move the complexity away from the front end.

Lastly, the route for enumerating leaders is no longer used by
mitoc.mit.edu: we can make this new route available only to logged-in
users (and start deprecating the old route).