Communicate when leader apps disabled
When applications for an activity are defined, but not enabled, we
should show activity chairs a warning about this and instructions on how
to enable new applications.

Similarly, when a participant does not have a pending application for an
activity and that activity is closed, we should show a brief
explanation.